Added check for a random hash before accepting account changes or deleting an account.

+++ public_html/usersettings.php
@@ -5,15 +5,15 @@
 // | Geeklog 1.3                                                               |
 // +---------------------------------------------------------------------------+
 // | usersettings.php                                                          |
+// |                                                                           |
 // | Geeklog user settings page.                                               |
-// |                                                                           |
 // +---------------------------------------------------------------------------+
-// | Copyright (C) 2000-2003 by the following authors:                         |
+// | Copyright (C) 2000-2004 by the following authors:                         |
 // |                                                                           |
-// | Authors: Tony Bibbs       - tony@tonybibbs.com                            |
-// |          Mark Limburg     - mlimburg@users.sourceforge.net                |
-// |          Jason Wittenburg - jwhitten@securitygeeks.com                    |
-// |          Dirk Haun        - dirk@haun-online.de                           |
+// | Authors: Tony Bibbs        - tony@tonybibbs.com                           |
+// |          Mark Limburg      - mlimburg@users.sourceforge.net               |
+// |          Jason Whittenburg - jwhitten@securitygeeks.com                   |
+// |          Dirk Haun         - dirk@haun-online.de                          |
 // +---------------------------------------------------------------------------+
 // |                                                                           |
 // | This program is free software; you can redistribute it and/or             |
@@ -32,7 +32,7 @@
 // |                                                                           |
 // +---------------------------------------------------------------------------+
 //
-// $Id: usersettings.php,v 1.70 2003/08/04 12:11:52 dhaun Exp $
+// $Id: usersettings.php,v 1.70.2.1 2004/01/23 10:11:02 dhaun Exp $
 
 include_once('lib-common.php');
 
@@ -145,9 +145,13 @@ function edituser()
     $result = DB_query("SELECT about,pgpkey FROM {$_TABLES['userinfo']} WHERE uid = {$_USER['uid']}");
     $A = DB_fetchArray($result);
 
+    $reqid = substr (md5 (uniqid (rand (), 1)), 1, 16);
+    DB_change ($_TABLES['users'], 'pwrequestid', "$reqid",
+                                  'username', $username);
+
     $preferences->set_var ('about_value', $A['about']);
     $preferences->set_var ('pgpkey_value', $A['pgpkey']);
-    $preferences->set_var ('uid_value', $_USER['uid']);
+    $preferences->set_var ('uid_value', $reqid);
     $preferences->set_var ('username_value', $_USER['username']);
 
     if ($_CONF['allow_account_delete'] == 1) {
@@ -157,7 +161,7 @@ function edituser()
         $preferences->set_var ('delete_text', $LANG04[95]);
         $preferences->set_var ('lang_button_delete', $LANG04[96]);
         $preferences->set_var ('delete_mode', 'confirmdelete');
-        $preferences->set_var ('account_id', $_USER['uid']);
+        $preferences->set_var ('account_id', $reqid);
         $preferences->parse ('delete_account_option', 'deleteaccount', false);
     } else {
         $preferences->set_var ('delete_account_option', '');
@@ -174,18 +178,22 @@ function edituser()
 /**
 * Ask user for confirmation to delete his/her account.
 *
-* @param    int      account_id   uid of account to delete (must match current user's uid)
+* @param    string   form_reqid   request id
 * @return   string   confirmation form
 *
 */
-function confirmAccountDelete ($account_id)
+function confirmAccountDelete ($form_reqid)
 {
-    global $_CONF, $_USER, $LANG04;
+    global $_CONF, $_TABLES, $_USER, $LANG04;
 
-    if ($account_id != $_USER['uid']) {
-        // now that doesn't look right - abort ...
+    if (DB_count ($_TABLES['users'], array ('pwrequestid', 'uid'), array ($form_reqid, $_USER['uid'])) != 1) {
+        // not found - abort
         return COM_refresh ($_CONF['site_url'] . '/index.php');
     }
+
+    $reqid = substr (md5 (uniqid (rand (), 1)), 1, 16);
+    DB_change ($_TABLES['users'], 'pwrequestid', "$reqid",
+                                  'uid', $_USER['uid']);
 
     $retval = '';
 
@@ -200,7 +208,7 @@ function confirmAccountDelete ($account_id)
     $confirm->set_var ('delete_text', $LANG04[95]);
     $confirm->set_var ('lang_button_delete', $LANG04[96]);
     $confirm->set_var ('delete_mode', 'deleteconfirmed');
-    $confirm->set_var ('account_id', $_USER['uid']);
+    $confirm->set_var ('account_id', $reqid);
 
     $retval .= COM_siteHeader ('menu');
     $retval .= COM_startBlock ($LANG04[97], '',
@@ -216,18 +224,20 @@ function confirmAccountDelete ($account_id)
 /**
 * Delete an account (keep in sync with delete_user() in admin/user.php).
 *
-* @param    uid      int   uid of account to delete
+* @param    string   form_reqid   request id
 * @return   string   redirection to main page (+ success msg)
 *
 */
-function deleteUserAccount ($uid)
+function deleteUserAccount ($form_reqid)
 {
     global $_CONF, $_TABLES, $_USER;
 
-    if ($uid != $_USER['uid']) {
-        // now that doesn't look right - abort ...
+    if (DB_count ($_TABLES['users'], array ('pwrequestid', 'uid'), array ($form_reqid, $_USER['uid'])) != 1) {
+        // not found - abort
         return COM_refresh ($_CONF['site_url'] . '/index.php');
     }
+
+    $uid = $_USER['uid'];
 
     // log the user out
     SESS_endUserSession ($_USER['uid']);
@@ -254,6 +264,9 @@ function deleteUserAccount ($uid)
     // avoid having orphand stories/comments by making them anonymous posts
     DB_query ("UPDATE {$_TABLES['comments']} SET uid = 1 WHERE uid = $uid");
     DB_query ("UPDATE {$_TABLES['stories']} SET uid = 1 WHERE uid = $uid");
+
+    // delete personal events
+    DB_delete ($_TABLES['personal_events'], 'uid', $uid);
 
     // now delete the user itself
     DB_delete ($_TABLES['users'], 'uid', $uid);
@@ -623,6 +636,15 @@ function saveuser($A)
         COM_errorLog('**** Inside saveuser in usersettings.php ****', 1);
     } 
 
+    $reqid = DB_getItem ($_TABLES['users'], 'pwrequestid',
+                         "uid = {$_USER['uid']}");
+    if ($reqid != $A['uid']) {
+        DB_change ($_TABLES['users'], 'pwrequestid', "NULL",
+                   'uid', $_USER['uid']);
+        COM_accessLog ("An attempt was made to illegally change the account information of user {$_USER['uid']}.");
+        return COM_refresh ($_CONF['site_url'] . '/index.php');
+    }
+
     if ($_CONF['allow_username_change'] == 1) {
         $A['new_username'] = strip_tags (COM_stripslashes ($A['new_username']));
         if (!empty ($A['new_username']) &&
@@ -895,10 +917,28 @@ if (!empty($_USER['username']) && !empty($mode)) {
                                  . '/usersettings.php?mode=preferences&msg=6');
         break;
     case 'confirmdelete':
-        $display .= confirmAccountDelete ($HTTP_POST_VARS['account_id']);
+        if (($_CONF['allow_account_delete'] == 1) && ($_USER['uid'] > 1)) {
+            if (isset ($HTTP_POST_VARS['account_id']) &&
+                    !empty ($HTTP_POST_VARS['account_id'])) {
+                $display .= confirmAccountDelete ($HTTP_POST_VARS['account_id']);
+            } else {
+                $display = COM_refresh ($_CONF['site_url'] . '/index.php');
+            }
+        } else {
+            $display = COM_refresh ($_CONF['site_url'] . '/index.php');
+        }
         break;
     case 'deleteconfirmed':
-        $display .= deleteUserAccount ($HTTP_POST_VARS['account_id']);
+        if (($_CONF['allow_account_delete'] == 1) && ($_USER['uid'] > 1)) {
+            if (isset ($HTTP_POST_VARS['account_id']) &&
+                    !empty ($HTTP_POST_VARS['account_id'])) {
+                $display .= deleteUserAccount ($HTTP_POST_VARS['account_id']);
+            } else {
+                $display = COM_refresh ($_CONF['site_url'] . '/index.php');
+            }
+        } else {
+            $display = COM_refresh ($_CONF['site_url'] . '/index.php');
+        }
         break;
     }
 } else {
